New PCI operation: get_dev_rom()
As in regions, the goal is to provide the client with
a way to get the base address and the size of the rom
without writing in the config file. This may be done by
reading the config file for the base address and getting
the size from the fs. But this is a better way.