Remove support for multiple host bridges.
Support was added based on the following description in
http://wiki.osdev.org/PCI :

"Start by checking if the device at bus 0, device 0 is a multi-function
device. If it's not a multi-function device, then there is only one PCI
host controller and bus 0, device 0, function 0 will be the PCI host
controller responsible for bus 0. If it is a multifunction device, then
bus 0, device 0, function 0 will be the PCI host controller responsible
for bus 0; bus 0, device 0, function 1 will be the PCI host controller
responsible for bus 1, etc (up to the number of functions supported)."

After a research, I haven't found any other reference to this in the Spec,
books or web sites. But I found some forums where people say this is wrong.